Hacked to use Floats instead of Doubles

+++ src/main/scala/dagger/core/FeatureFunctionFactory.scala
@@ -7,5 +7,5 @@ abstract class FeatureFunctionFactory[D: ClassTag, S <: TransitionState: ClassTa
 }
 
 abstract class FeatureFunction[D: ClassTag, S <: TransitionState: ClassTag, A <: TransitionAction[S]: ClassTag] {
-  def features(data: D, state: S, action: A): Map[Int, Double]
+  def features(data: D, state: S, action: A): Map[Int, Float]
 }
