+++ src/main/scala/dagger/ml/ClassificationInstance.scala
@@ -10,7 +10,7 @@ import scala.reflect.ClassTag
  * Time: 2:43 PM
  */
 
-case class Instance[T](feats: List[Map[Int, Double]], labels: Array[T], weightLabels: Array[T], costs: Array[Double] = null) {
+case class Instance[T](feats: List[Map[Int, Float]], labels: Array[T], weightLabels: Array[T], costs: Array[Float] = null) {
 
   def featureVector = feats
 
@@ -41,7 +41,7 @@ case class Instance[T](feats: List[Map[Int, Double]], labels: Array[T], weightLa
 
 object Instance {
 
-  def construct[T: ClassTag](feats: List[Map[Int, Double]], ilabels: Array[T], icosts: Array[Double], correct: Array[Boolean]): Instance[T] = {
+  def construct[T: ClassTag](feats: List[Map[Int, Float]], ilabels: Array[T], icosts: Array[Float], correct: Array[Boolean]): Instance[T] = {
     assert(ilabels.size > 1 && icosts.size > 1, "Insufficient costs and labels (<1) for Instance.")
     val scosts = (ilabels, icosts, feats).zipped.toList.sortBy(_._2)
     var (maxCost, minCost) = (scosts.head._2, scosts.last._2)
@@ -51,13 +51,13 @@ object Instance {
   def fromSerialString[T](str: String): Instance[T] = {
     val lines = str.split("\n")
     val number = lines(0).toInt
-    val feats = ((1 to number) map (i => lines(i).split(" ").view.map(_.toInt -> 1.0).toMap)).toList
+    val feats = ((1 to number) map (i => lines(i).split(" ").view.map(_.toInt -> 1.0f).toMap)).toList
     val labels = lines(number+1).split(" ").map {
       _ match {
         case _ => "Unserializing the following action is unsupported: " + _
       }
     }.asInstanceOf[Array[T]]
-    val costs = lines(number+2).split(" ").map(_.toDouble)
+    val costs = lines(number+2).split(" ").map(_.toFloat)
     new Instance[T](feats, labels, labels, costs)
   }
 }
